Activity indicator для кнопки вынесен в отдельную структуру. Данный индикатор отображается при нажатии на кнопки, которые инициируют сетевые вызовы.
